+++ pyprocessmacro/models.py
@@ -16,6 +16,7 @@ from .utils import (
     eval_expression,
     bias_corrected_ci,
     z_score,
+    t_score,
     percentile_ci,
     find_significance_region,
 )
@@ -316,13 +317,13 @@ class OLSOutcomeModel(BaseOutcomeModel):
         t = betas / se
         p = stats.t.sf(np.abs(t), df_e) * 2
         conf = self._options["conf"]
-        zscore = z_score(conf)
+        tcrit = t_score(conf, df_e)  # OLS intervals use the t distribution, as PROCESS does (#40)
         R2 = 1 - resid.var() / y.var()
-        adjR2 = 1 - (1 - R2) * ((n_obs - 1) / (n_obs - n_vars - 1))
+        adjR2 = 1 - (1 - R2) * ((n_obs - 1) / df_e)  # n_vars already counts the constant (#41)
         F = (R2 / df_r) / ((1 - R2) / df_e)
-        F_pval = 1 - stats.f.cdf(F, df_r, df_e)
-        llci = betas - (se * zscore)
-        ulci = betas + (se * zscore)
+        F_pval = stats.f.sf(F, df_r, df_e)
+        llci = betas - (se * tcrit)
+        ulci = betas + (se * tcrit)
         names = [self._symb_to_var.get(x, x) for x in self._exogvars]
         estimation_results = {
             "betas": betas,
@@ -418,19 +419,19 @@ class LogitOutcomeModel(BaseOutcomeModel, BaseLogit):
 
         # GOF statistics
         llmodel = self._loglike(betas)
-        lmodel = np.exp(llmodel)
         minus2ll = -2 * llmodel
 
         null_model = NullLogitModel(self._endog, self._options)
         betas_null = null_model._optimize()
         llnull = null_model._loglike(betas_null)
-        lnull = np.exp(llnull)
 
         d = 2 * (llmodel - llnull)
         pvalue = stats.chi2.sf(d, self._n_vars - 1)
         mcfadden = 1 - llmodel / llnull
-        coxsnell = 1 - (lnull / lmodel) ** (2 / self._n_obs)
-        nagelkerke = coxsnell / (1 - lnull ** (2 / self._n_obs))
+        # Likelihood ratios are taken in log space: exp(llnull) underflows to 0 beyond about a
+        # thousand observations, which turned both pseudo R-squared into NaN (#42).
+        coxsnell = 1 - np.exp(2 * (llnull - llmodel) / self._n_obs)
+        nagelkerke = coxsnell / (1 - np.exp(2 * llnull / self._n_obs))
         names = [self._symb_to_var.get(x, x) for x in self._exogvars]
         estimation_results = {
             "betas": betas,
@@ -1585,9 +1586,12 @@ class DirectEffectModel(object):
             dot(grad, vcv), np.transpose(grad)
         )  # V(Grad(X)) = Grad(X).V(X).Grad'(X)
         se = np.sqrt(var)
-        zscore = z_score(conf)
-        llci = betas - (se * zscore)
-        ulci = betas + (se * zscore)
+        if self._is_logit:
+            crit = z_score(conf)
+        else:  # OLS intervals use the t distribution, as PROCESS does (#40)
+            crit = t_score(conf, self._model.estimation_results["df_e"])
+        llci = betas - (se * crit)
+        ulci = betas + (se * crit)
         return betas, se, llci, ulci
 
     def coeff_summary(self):